move LocalPaperDirectory
